Add files via upload
I added portal sprites and started the code on them to transport the fairies to the next room. I need to add the code where the player needs to have a key but I am going to wait until I figure out how to get the player to go into the next room. The code is working like it will go into the room when the player hits the portal, but the player is not going to the next room.  I Also got started on the background for room2 and some tiles.