Lvl 1 background
I redid the level one background. I also added code to the player so that whenever the player hits the object it takes away health. I also add code that makes it be invincible for a few seconds so that the player doesn't lose health for the same hit over and over. I tried to do a kickback for whenever the player hits the object but i could not get it to not be jerky so I took it out.